requirements.txt
#install flask

flask
flask-sqlalchemy
flask-migrate
flask-cors